a bunch of systemd service files
This should make life much easier to automate stuff. I used a chatgpt
with some manual edits, so please check it out also yourself but it
all works for me.

The nym-client there is twice just if there is a need to run two
instances on one box. I kept the README port 1978 for that service.

Cheers.